Split core preludes into their own files
